requirements.txt
requests
python-dotenv
logzero
